added walking and reload sound effects

+++ Assets/ScriptableObjects/Guns/Gun.cs
@@ -5,7 +5,7 @@ public class Gun : Pickeable
 {
     public int ShootDamage;
     public int ShotDistance;
-    public int Recoil;
+    public float Recoil;
     public int BulletsInMagazine;
     public GameObject GunPrefab;
 }
